pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gra-shnoll"
version = "1.0.0"
description = "GRA-Shnoll cosmophysical navigation engine for drones"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "astropy>=5.3",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "tqdm>=4.65"
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "black>=23.7", "mypy>=1.5"]
